:zap: Improve Member Service

See merge request s08-webmobile1-sub2/S08P12B104!101

+++ server/src/main/java/com/ssafy/live/account/realtor/service/RealtorService.java
@@ -26,6 +26,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -65,20 +66,8 @@ public class RealtorService {
             return response.fail("이미 회원가입된 중개사번호입니다.", HttpStatus.BAD_REQUEST);
         }
         String imgSrc = s3Service.upload(file);
-        Realtor realtor = Realtor.builder()
-            .password(passwordEncoder.encode(signUp.getPassword()))
-            .name(signUp.getName())
-            .email(signUp.getEmail())
-            .phone(signUp.getPhone())
-            .corp(signUp.getCorp())
-            .description(signUp.getDescription())
-            .businessNumber(signUp.getBusinessNumber())
-            .businessAddress(signUp.getBusinessAddress())
-            .startDate(signUp.getStartDate())
-            .registrationNumber(signUp.getRegistrationNumber())
-            .imageSrc(imgSrc)
-            .roles(Collections.singletonList(Authority.ROLE_REALTOR.name()))
-            .build();
+        Realtor realtor = RealtorRequest.SignUp.toEntity(signUp, imgSrc, passwordEncoder.encode(signUp.getPassword()));
+
         realtorRepository.save(realtor);
 
         return response.success("회원가입에 성공했습니다.");
@@ -93,36 +82,12 @@ public class RealtorService {
         Authentication authentication = authenticationManager.authenticate(authenticationToken);
         SecurityContextHolder.getContext().setAuthentication(authentication);
         CommonResponse.TokenInfo tokenInfo = jwtTokenProvider.generateToken(authentication);
-
         redisTemplate.opsForValue()
             .set("RT:" + authentication.getName(), tokenInfo.getRefreshToken(), tokenInfo.getRefreshTokenExpirationTime(), TimeUnit.MILLISECONDS);
 
         return response.success(tokenInfo, "로그인에 성공했습니다.", HttpStatus.OK);
     }
 
-    public ResponseEntity<?> reissue(RealtorRequest.Reissue reissue) {
-        if (!jwtTokenProvider.validateToken(reissue.getRefreshToken())) {
-            return response.fail("Refresh Token 정보가 유효하지 않습니다.", HttpStatus.BAD_REQUEST);
-        }
-
-        Authentication authentication = jwtTokenProvider.getAuthentication(reissue.getAccessToken());
-        String refreshToken = (String)redisTemplate.opsForValue().get("RT:" + authentication.getName());
-
-        if(ObjectUtils.isEmpty(refreshToken)) {
-            return response.fail("잘못된 요청입니다.", HttpStatus.BAD_REQUEST);
-        }
-
-        if(!refreshToken.equals(reissue.getRefreshToken())) {
-            return response.fail("Refresh Token 정보가 일치하지 않습니다.", HttpStatus.BAD_REQUEST);
-        }
-
-        CommonResponse.TokenInfo tokenInfo = jwtTokenProvider.generateToken(authentication);
-
-        redisTemplate.opsForValue()
-            .set("RT:" + authentication.getName(), tokenInfo.getRefreshToken(), tokenInfo.getRefreshTokenExpirationTime(), TimeUnit.MILLISECONDS);
-
-        return response.success(tokenInfo, "Token 정보가 갱신되었습니다.", HttpStatus.OK);
-    }
 
     public ResponseEntity<?> logout(RealtorRequest.Logout logout) {
         Authentication authentication = jwtTokenProvider.getAuthentication(logout.getAccessToken());
@@ -148,19 +113,6 @@ public class RealtorService {
         return response.fail("해당하는 회원을 찾을 수 없습니다.", HttpStatus.BAD_REQUEST);
     }
 
-    public ResponseEntity<?> authority() {
-        // SecurityContext에 담겨 있는 authentication userId 정보
-        String userId = SecurityUtil.getCurrentUserId();
-
-        Realtor realtor = realtorRepository.findByBusinessNumber(userId)
-            .orElseThrow(() -> new UsernameNotFoundException("No authentication information."));
-
-        // add ROLE_USER
-        realtor.getRoles().add(Authority.ROLE_REALTOR.name());
-        realtorRepository.save(realtor);
-
-        return response.success();
-    }
 
     public ResponseEntity<?> findRealtorDetail(Long realtorNo) {
         Realtor realtor = realtorRepository.findById(realtorNo).orElseThrow(()->new BadRequestException(REALTOR_NOT_FOUND));
@@ -180,8 +132,11 @@ public class RealtorService {
     }
 
     @Transactional
-    public ResponseEntity<?> updateRealtor(Long realtorNo, RealtorRequest.Update request, MultipartFile file) throws IOException {
-        Realtor realtor = realtorRepository.findById(realtorNo).orElseThrow(()->new BadRequestException(REALTOR_NOT_FOUND));
+    public ResponseEntity<?> updateRealtor(UserDetails realtors, RealtorRequest.Update request, MultipartFile file) throws IOException {
+        Realtor realtor = realtorRepository.findByBusinessNumber(realtors.getUsername()).get();
+        if(realtor == null) {
+            return response.fail("해당하는 공인중개사를 찾을 수 없습니다.", HttpStatus.BAD_REQUEST);
+        }
         String preImg = realtor.getImageSrc();
         if(file != null) {
             s3Service.deleteFile(preImg);
@@ -189,13 +144,8 @@ public class RealtorService {
         String imgSrc = s3Service.upload(file);
         realtor.updateRealtor(request, passwordEncoder.encode(request.getPassword()), imgSrc);
         realtorRepository.save(realtor);
-        RealtorResponse.Update updateRealtor = RealtorResponse.Update.builder()
-            .phone(realtor.getPhone())
-            .email(realtor.getEmail())
-            .description(realtor.getDescription())
-            .imageSrc(realtor.getImageSrc())
-            .build();
-        return response.success(updateRealtor, "공인중개사 정보 수정을 완료했습니다.", HttpStatus.OK);
+
+        return response.success(RealtorResponse.UpdateRealtor.toEntity(realtor), "공인중개사 정보 수정을 완료했습니다.", HttpStatus.OK);
     }
 
     public ResponseEntity<?> temporaryPassword(RealtorRequest.FindPassword request) {
@@ -233,5 +183,43 @@ public class RealtorService {
             findRealtors = realtorRepository.findAllByOrderByCountByItemDesc();
         }
         return response.success(findRealtors,"메인페이지의 공인중개사 목록을 조회하였습니다.", HttpStatus.OK);
+    }
+
+    public ResponseEntity<?> reissue(RealtorRequest.Reissue reissue) {
+        if (!jwtTokenProvider.validateToken(reissue.getRefreshToken())) {
+            return response.fail("Refresh Token 정보가 유효하지 않습니다.", HttpStatus.BAD_REQUEST);
+        }
+
+        Authentication authentication = jwtTokenProvider.getAuthentication(reissue.getAccessToken());
+        String refreshToken = (String)redisTemplate.opsForValue().get("RT:" + authentication.getName());
+
+        if(ObjectUtils.isEmpty(refreshToken)) {
+            return response.fail("잘못된 요청입니다.", HttpStatus.BAD_REQUEST);
+        }
+
+        if(!refreshToken.equals(reissue.getRefreshToken())) {
+            return response.fail("Refresh Token 정보가 일치하지 않습니다.", HttpStatus.BAD_REQUEST);
+        }
+
+        CommonResponse.TokenInfo tokenInfo = jwtTokenProvider.generateToken(authentication);
+
+        redisTemplate.opsForValue()
+            .set("RT:" + authentication.getName(), tokenInfo.getRefreshToken(), tokenInfo.getRefreshTokenExpirationTime(), TimeUnit.MILLISECONDS);
+
+        return response.success(tokenInfo, "Token 정보가 갱신되었습니다.", HttpStatus.OK);
+    }
+
+    public ResponseEntity<?> authority() {
+        // SecurityContext에 담겨 있는 authentication userId 정보
+        String userId = SecurityUtil.getCurrentUserId();
+
+        Realtor realtor = realtorRepository.findByBusinessNumber(userId)
+            .orElseThrow(() -> new UsernameNotFoundException("No authentication information."));
+
+        // add ROLE_USER
+        realtor.getRoles().add(Authority.REALTOR.name());
+        realtorRepository.save(realtor);
+
+        return response.success();
     }
 }
